LayoutTests/inspector/profiler/detailed-heapshots-summary.html

<html>
<head>
  <script src="../../http/tests/inspector/inspector-test.js"></script>
  <script src="detailed-heapshots-test.js"></script>
<script>

function test()
{
    var instanceCount = 555;
    function createHeapSnapshot()
    {
        return InspectorTest.createHeapSnapshot(instanceCount);
    }

    InspectorTest.runDetailedHeapshotTestSuite([
        function testSorting(next)
        {
            InspectorTest.takeAndOpenSnapshot(createHeapSnapshot, step1);

            function step1()
            {
                InspectorTest.switchToView("Summary", step2);
            }

            var columns;
            var currentColumn;
            var currentColumnOrder;

            function step2()
            {
                columns = InspectorTest.viewColumns();
                currentColumn = 0;
                currentColumnOrder = false;
                step3();
            }

            function step3()
            {
                if (currentColumn >= columns.length) {
                    next();
                    return;
                }

                InspectorTest.clickColumn(columns[currentColumn], step4);
            }

            function step4(newColumnState)
            {
                columns[currentColumn] = newColumnState;
                var contents = InspectorTest.columnContents(columns[currentColumn]);
                InspectorTest.assertEquals(true, !!contents.length, "column contents");
                var sortTypes = { object: "text", count: "number", shallowSize: "size", retainedSize: "size" };
                InspectorTest.assertEquals(true, !!sortTypes[columns[currentColumn].identifier], "sort by identifier");
                InspectorTest.checkArrayIsSorted(contents, sortTypes[columns[currentColumn].identifier], columns[currentColumn].sort);

                if (!currentColumnOrder)
                    currentColumnOrder = true;
                else {
                    currentColumnOrder = false;
                    ++currentColumn;
                }
                step3();
            }
        },

        function testShowNext(next)
        {
            InspectorTest.takeAndOpenSnapshot(createHeapSnapshot, step1);

            function step1()
            {
                InspectorTest.switchToView("Summary", step2);
            }

            function step2()
            {
                var row = InspectorTest.findRow("object", "A");
                InspectorTest.assertEquals(true, !!row, "\"A\" row");
                InspectorTest.expandRow(row, step3);
            }

            function step3(row)
            {
                var rowsShown = InspectorTest.countDataRows(row);
                InspectorTest.assertEquals(true, rowsShown <= instanceCount, "shown more instances than created: " + rowsShown + " > " + instanceCount);
                if (rowsShown < instanceCount) {
                    var buttonsNode = InspectorTest.findButtonsNode(row);
                    InspectorTest.assertEquals(true, !!buttonsNode, "buttons node");
                    InspectorTest.clickShowMoreButton("showNext", buttonsNode, step3);
                } else if (rowsShown === instanceCount) {
                    var buttonsNode = InspectorTest.findButtonsNode(row);
                    InspectorTest.assertEquals(false, !!buttonsNode, "buttons node found when all instances are shown!");
                    next();
                }
            }
        },

        function testShowAll(next)
        {
            InspectorTest.takeAndOpenSnapshot(createHeapSnapshot, step1);

            function step1()
            {
                InspectorTest.switchToView("Summary", step2);
            }

            function step2()
            {
                var row = InspectorTest.findRow("object", "A");
                InspectorTest.assertEquals(true, !!row, "\"A\" row");
                InspectorTest.expandRow(row, step3);
            }

            function step3(row)
            {
                var count = row.data["count"];
                InspectorTest.assertEquals(instanceCount, count);
                var buttonsNode = InspectorTest.findButtonsNode(row);
                InspectorTest.assertEquals(true, !!buttonsNode, "buttons node");
                var words = buttonsNode.showAll.textContent.split(" ");
                for (var i = 0; i < words.length; ++i) {
                    var maybeNumber = parseInt(words[i], 10);
                    if (!isNaN(maybeNumber))
                        InspectorTest.assertEquals(instanceCount, maybeNumber, buttonsNode.showAll.textContent);
                }
                InspectorTest.clickShowMoreButton("showAll", buttonsNode, step4);
            }

            function step4(row)
            {
                var rowsShown = InspectorTest.countDataRows(row);
                InspectorTest.assertEquals(instanceCount, rowsShown, "after showAll click");
                var buttonsNode = InspectorTest.findButtonsNode(row);
                InspectorTest.assertEquals(false, !!buttonsNode, "buttons node found when all instances are shown!");
                next();
            }
        },

        function testExpansionPreservedWhenSorting(next)
        {
            InspectorTest.takeAndOpenSnapshot(createHeapSnapshot, step1);

            function step1()
            {
                InspectorTest.switchToView("Summary", step2);
            }

            var columns;
            function step2()
            {
                columns = InspectorTest.viewColumns();
                InspectorTest.clickColumn(columns[0], step3);
            }

            function step3()
            {
                var row = InspectorTest.findRow("object", "B");
                InspectorTest.assertEquals(true, !!row, "\"B\" row");
                InspectorTest.expandRow(row, expandBInstance);
                function expandBInstance(row)
                {
                    bInstanceRow = row.children[0];
                    InspectorTest.assertEquals(true, !!bInstanceRow, "\"B\" instance row");
                    InspectorTest.expandRow(bInstanceRow, expandA);
                }
                function expandA(row)
                {
                    function propertyMatcher(data)
                    {
                        return data.name === "a" && data.value.charAt(0) === "A";
                    }
                    var aRow = InspectorTest.findRow("object", propertyMatcher, row);
                    InspectorTest.assertEquals(true, !!aRow, "\"a: A\" row");
                    InspectorTest.expandRow(aRow, step4);
                }
            }

            var columnContents;
            var testStates = ["", "expanded", "show more", "show all"];
            function step4()
            {
                testStates.shift();
                if (!testStates.length) {
                    next();
                    return;
                }

                columnContents = InspectorTest.columnContents(columns[0]);
                InspectorTest.clickColumn(columns[0], clickTwice);
                function clickTwice()
                {
                    InspectorTest.clickColumn(columns[0], step5);
                }
            }

            function step5()
            {
                var newColumnContents = InspectorTest.columnContents(columns[0]);
                InspectorTest.assertColumnContentsEqual(columnContents, newColumnContents);

                var row = InspectorTest.findRow("object", "B");
                InspectorTest.assertEquals(true, !!row, "\"B\" row");
                var buttonsNode = InspectorTest.findButtonsNode(row);
                if (testStates[0] === "expanded" || testStates[0] === "show more") {
                    InspectorTest.assertEquals(true, !!buttonsNode, testStates[0] + ": no buttons node found!");
                    if (testStates[0] === "expanded")
                        InspectorTest.clickShowMoreButton("showNext", buttonsNode, step4);
                    else
                        InspectorTest.clickShowMoreButton("showAll", buttonsNode, step4);
                } else {
                    InspectorTest.assertEquals(false, !!buttonsNode, "buttons node found when all instances are shown!");
                    step4();
                }
            }
        }
    ]);
}

</script>
</head>
<body onload="runTest()">
<p>
Tests Summary view of detailed heap snapshots.
</p>
</body>
</html>
