LayoutTests/platform/chromium/inspector/profiler/detailed-heapshots-summary-expected.txt

Tests Summary view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testSorting

Running: testShowNext

Running: testShowAll

Running: testExpansionPreservedWhenSorting

Profiler was disabled.
